projdocker/requirements.txt
aniso8601==9.0.1
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cryptography==36.0.2
fastapi==0.75.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
google-api-core==2.7.1
google-api-python-client==2.41.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
graphene==3.0
graphql-core==3.1.5
graphql-relay==3.1.5
greenlet==1.1.2
h11==0.13.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
pika==1.2.0
promise==2.3
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.7
python-gql==0.2.4
requests==2.27.1
rsa==4.8
Rx==1.6.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.17.1
starlette-graphene3==0.5.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.17.6
Werkzeug==2.0.2
